Add boa and boa-plot entry points to pyproject.toml
Allow users to start BOA with `boa -c config` or `pythong -m boa -c config`
and start plotting with `boa-plot -sp scheduler` or `pythong -m boa.plot -sp scheduler`
