+++ boa/runner.py
@@ -9,6 +9,7 @@ Runner that calls your :mod:`.wrappers` to run your model and poll the trial sta
 
 import concurrent.futures
 import logging
+import multiprocessing
 from collections import defaultdict
 from typing import Any, Dict, Iterable, Set
 
@@ -16,7 +17,7 @@ from ax.core.base_trial import TrialStatus
 from ax.core.runner import Runner
 from ax.core.trial import Trial
 
-from boa.logger import get_logger, queue
+from boa.logger import get_logger
 from boa.metaclasses import RunnerRegister
 from boa.utils import serialize_init_args
 from boa.wrappers.base_wrapper import BaseWrapper
@@ -28,6 +29,7 @@ class WrappedJobRunner(Runner, metaclass=RunnerRegister):
     def __init__(self, wrapper: BaseWrapper = None, *args, **kwargs):
 
         self.wrapper = wrapper or BaseWrapper()
+        self.queue = multiprocessing.Manager().Queue()
         super().__init__(*args, **kwargs)
 
     def run(self, trial: Trial) -> Dict[str, Any]:
@@ -42,7 +44,7 @@ class WrappedJobRunner(Runner, metaclass=RunnerRegister):
         Returns:
             Dict of run metadata from the deployment process.
         """
-        qh = logging.handlers.QueueHandler(queue)
+        qh = logging.handlers.QueueHandler(self.queue)
         logger = logging.getLogger()
         logger.addHandler(qh)
         ax_logger = logging.getLogger("ax")
@@ -119,7 +121,7 @@ class WrappedJobRunner(Runner, metaclass=RunnerRegister):
 
         parents = self.__class__.mro()[1:]  # index 0 is the class itself
 
-        properties = serialize_init_args(self, parents=parents, match_private=True, exclude_fields=["wrapper"])
+        properties = serialize_init_args(self, parents=parents, match_private=True, exclude_fields=["wrapper", "queue"])
 
         properties["__type"] = self.__class__.__name__
         return properties
